requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
simple-salesforce==1.12.6
pydantic==2.8.2
pydantic-settings==2.3.4
python-dotenv==1.0.1
httpx==0.27.0
pytest==8.2.2
pytest-asyncio==0.23.7
openai==1.37.0
python-jobspy==1.1.52
jinja2==3.1.4
